had to fix everything
